phase-2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
pydantic==2.6.4
pydantic-settings==2.2.1
aiosqlite==0.20.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.12
better-exceptions==0.3.3